+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/schedule/AiringScheduleScreen.kt
@@ -230,6 +230,7 @@ object AiringScheduleScreen {
                                                         airingAt = it.airingAt,
                                                     )
                                                 },
+                                                showDate = false,
                                                 colorCalculationState = colorCalculationState,
                                             )
                                         }