+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/schedule/AiringScheduleSortOption.kt
@@ -8,9 +8,9 @@ import com.thekeeperofpie.artistalleydatabase.compose.filter.SortOption
 enum class AiringScheduleSortOption(@StringRes override val textRes: Int): SortOption {
 
     POPULARITY(R.string.anime_airing_schedule_sort_popularity),
-    ID(R.string.anime_airing_schedule_sort_id),
     TIME(R.string.anime_airing_schedule_sort_time),
     EPISODE(R.string.anime_airing_schedule_sort_episode),
+    ID(R.string.anime_airing_schedule_sort_id),
     ;
 
     fun toApiValue(ascending: Boolean) = when(this) {